+++ app/Actions/Helpers/Tag/DeleteTag.php
@@ -8,25 +8,18 @@
  *
 */
 
-namespace App\Actions\Catalogue\Tag;
+namespace App\Actions\Helpers\Tag;
 
 use App\Actions\OrgAction;
-use App\Models\Catalogue\Tag;
+use App\Models\Helpers\Tag;
 use Lorisleiva\Actions\ActionRequest;
 
-class UpdateTag extends OrgAction
+class DeleteTag extends OrgAction
 {
-    public function handle(Tag $tag, array $modelData): Tag
+    public function handle(Tag $tag): Tag
     {
-        $tag->update($modelData);
+        $tag->delete();
         return $tag;
-    }
-
-    public function rules(): array
-    {
-        return [
-            'name' => ['sometimes', 'nullable', 'string', 'max:255'],
-        ];
     }
 
     public function asController(Tag $tag, ActionRequest $request)
