+++ app/Models/Catalogue/Product.php
@@ -16,6 +16,7 @@ use App\Models\CRM\BackInStockReminder;
 use App\Models\CRM\Favourite;
 use App\Models\Dropshipping\Portfolio;
 use App\Models\Goods\TradeUnit;
+use App\Models\Helpers\Tag;
 use App\Models\Inventory\OrgStock;
 use App\Models\SysAdmin\Group;
 use App\Models\SysAdmin\Organisation;
@@ -23,6 +24,7 @@ use App\Models\Traits\HasHistory;
 use App\Models\Traits\HasImage;
 use App\Models\Traits\HasUniversalSearch;
 use App\Models\Web\Webpage;
+use Illuminate\Database\Eloquent\Collection as LaravelCollection;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -36,7 +38,6 @@ use OwenIt\Auditing\Contracts\Auditable;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\Sluggable\HasSlug;
 use Spatie\Sluggable\SlugOptions;
-use Illuminate\Database\Eloquent\Collection as LaravelCollection;
 
 /**
  *
@@ -83,28 +84,28 @@ use Illuminate\Database\Eloquent\Collection as LaravelCollection;
  * @property string|null $historic_source_id
  * @property bool $is_for_sale For sale products including out of stock
  * @property-read \App\Models\Catalogue\Asset|null $asset
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Helpers\Audit> $audits
- * @property-read \Illuminate\Database\Eloquent\Collection<int, BackInStockReminder> $backInStockReminders
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Catalogue\Collection> $collections
+ * @property-read LaravelCollection<int, \App\Models\Helpers\Audit> $audits
+ * @property-read LaravelCollection<int, BackInStockReminder> $backInStockReminders
+ * @property-read LaravelCollection<int, \App\Models\Catalogue\Collection> $collections
  * @property-read \App\Models\Helpers\Currency $currency
  * @property-read \App\Models\Catalogue\ProductCategory|null $department
  * @property-read \App\Models\Catalogue\ProductCategory|null $family
- * @property-read \Illuminate\Database\Eloquent\Collection<int, Favourite> $favourites
+ * @property-read LaravelCollection<int, Favourite> $favourites
  * @property-read Group $group
  * @property-read \App\Models\Catalogue\HistoricAsset|null $historicAsset
- * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Catalogue\HistoricAsset> $historicAssets
+ * @property-read LaravelCollection<int, \App\Models\Catalogue\HistoricAsset> $historicAssets
  * @property-read \App\Models\Helpers\Media|null $image
  * @property-read \Spatie\MediaLibrary\MediaCollections\Models\Collections\MediaCollection<int, \App\Models\Helpers\Media> $images
  * @property-read Product|null $mainProduct
  * @property-read \Spatie\MediaLibrary\MediaCollections\Models\Collections\MediaCollection<int, \App\Models\Helpers\Media> $media
- * @property-read \Illuminate\Database\Eloquent\Collection<int, OrgStock> $orgStocks
+ * @property-read LaravelCollection<int, OrgStock> $orgStocks
  * @property-read Organisation $organisation
  * @property-read Portfolio|null $portfolio
- * @property-read \Illuminate\Database\Eloquent\Collection<int, Product> $productVariants
+ * @property-read LaravelCollection<int, Product> $productVariants
  * @property-read \App\Models\Catalogue\Shop|null $shop
  * @property-read \App\Models\Catalogue\ProductStats|null $stats
  * @property-read \App\Models\Catalogue\ProductCategory|null $subDepartment
- * @property-read \Illuminate\Database\Eloquent\Collection<int, TradeUnit> $tradeUnits
+ * @property-read LaravelCollection<int, TradeUnit> $tradeUnits
  * @property-read \App\Models\Helpers\UniversalSearch|null $universalSearch
  * @property-read Webpage|null $webpage
  * @method static \Database\Factories\Catalogue\ProductFactory factory($count = null, $state = [])
