+++ app/Models/Helpers/Tag.php
@@ -8,7 +8,7 @@
  *
 */
 
-namespace App\Models\Catalogue;
+namespace App\Models\Helpers;
 
 use App\Models\Goods\TradeUnit;
 use Illuminate\Database\Eloquent\Model;
@@ -16,6 +16,26 @@ use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Spatie\Sluggable\HasSlug;
 use Spatie\Sluggable\SlugOptions;
 
+/**
+ *
+ *
+ * @property int $id
+ * @property int $group_id
+ * @property int $organisation_id
+ * @property int|null $shop_id
+ * @property string $slug
+ * @property string $name
+ * @property string $scope
+ * @property array<array-key, mixed> $data
+ * @property int $number_models
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, TradeUnit> $tradeUnits
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Tag newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Tag newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|Tag query()
+ * @mixin \Eloquent
+ */
 class Tag extends Model
 {
     use HasSlug;
