+++ routes/grp/web/models.php
@@ -32,10 +32,6 @@ use App\Actions\Catalogue\ProductCategory\StoreSubDepartment;
 use App\Actions\Catalogue\ProductCategory\UpdateProductCategory;
 use App\Actions\Catalogue\Shop\StoreShop;
 use App\Actions\Catalogue\Shop\UpdateShop;
-use App\Actions\Catalogue\Tag\AttachTagsToModel;
-use App\Actions\Catalogue\Tag\DetachTagFromModel;
-use App\Actions\Catalogue\Tag\StoreTag;
-use App\Actions\Catalogue\Tag\UpdateTag;
 use App\Actions\Comms\Email\PublishEmail;
 use App\Actions\Comms\Email\UpdateEmailUnpublishedSnapshot;
 use App\Actions\Comms\EmailTemplate\UpdateEmailTemplate;
@@ -151,6 +147,10 @@ use App\Actions\Helpers\GoogleDrive\AuthorizeClientGoogleDrive;
 use App\Actions\Helpers\GoogleDrive\CallbackClientGoogleDrive;
 use App\Actions\Helpers\Media\AttachAttachmentToModel;
 use App\Actions\Helpers\Media\DetachAttachmentFromModel;
+use App\Actions\Helpers\Tag\AttachTagsToModel;
+use App\Actions\Helpers\Tag\DetachTagFromModel;
+use App\Actions\Helpers\Tag\StoreTag;
+use App\Actions\Helpers\Tag\UpdateTag;
 use App\Actions\HumanResources\ClockingMachine\DeleteClockingMachine;
 use App\Actions\HumanResources\ClockingMachine\StoreClockingMachine;
 use App\Actions\HumanResources\ClockingMachine\UpdateClockingMachine;
